xenmgr: Modify the templates for HVM NDVM
Switch NDVM to running as a HVM.

Bump up the RAM to avoid OOM.

qemu-dm-path needs to be specified to allow an HVM to boot.

kernel-extract must be unset otherwise we fail extracting a kernel as we
can't mount the vhd.

We also need to specify boot so qemu looks for boot devices.

Also specify a serial port so stubdom can output the logs.  We use file:
even though it is a fifo/pipe since we only want to output through the
serial port.  Lock-ups in the guest/qemu were observed with pipe: on a
single filename along with a getty in the guest on that serial line.

Specify nogfx to disable graphics.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>