v4v: Fix SMAP & HVM guest use
When SMAP is enabled, v4v_send fails from an HVM domain.

An HVM userspace virtual address is passed in as the "from" address, and
the destination is a previously registered v4v ring.  When
v4v_memcpy_to_guest_ring_from_guest is called, we are in the HVM's vcpu.

The call chains is something like:
v4v_memcpy_to_guest_ring_from_guest
  copy_from_guest
   ..
    __hvm_copy
      paging_gva_to_gfn
       ..
        guest_walk_tables

And guest_walk_tables fails with an emulated SMAP violation.  Disabling
the smap_check allows the page lookup to succeed.

It could that a smap violation occurs as the vcpu is in kernel mode when
it issues the hypercall, but our input buffer is a userspace address.

We don't need to play any stac/clac games ince the failure is just in
the address lookup.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>