input-daemon: Only run at runlevel 5
input-daemon is only needed when running surfman, so start it at the
same runlevel - 5.

Note, this means there is log spam about the missing DBus endpoint
com.citrix.xenclient.input when in runlevel 3.  However, without this,
running telinit 3 from telinit 5 means there is no input to dom0 when we
drop to tty1.  A better way would be to call inputserver to ungrab input
for switch focus.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>