dbus: Remove v4v.patch
DEPENDS in the dbus recipe did not pull in the libv4v.h header, so v4v
support would be off in dbus.

With the converison of syncvm to use INET_IS_V4V=1, there are no users
of family=v4 for DBUS_SYSTEM_BUS_ADDRESS.  Hence, we can just drop the
patch.

OXT-1306

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>