qemu-dm: 4.1 Add crypto-Make-getrandom-non-blocking.patch
getrandom() blocks by default.  In a xen stubdom, there is a lack of
entropy during boot, so this call hangs indefinitely.  The xen toolstack
doesn't see qemu identify itself as running and aborts the VM startup.

Switch to calling getrandom() as non-blocking during initialization
check.  In that case, an unitialized entropy pool with return eith
EAGAIN.  For qcrypto_random_init, that means we'll fall back to
/dev/urandom which warns to dmesg while still providing random data.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>