xen: Patch block-tap script
This brings in a patched block-tap script.  It provides the device
sharing checks previously done in libxl.

Calling tap-ctl passes an open file descriptor to tapback which messes
up locking.  That is why we have `eval "exec $_lockfd>&-` to close the
flock fd.

We also play games for the keydir.  This tries to mimic what's in xenmgr
already... I guess we could let xenmgr set those variables before
execing xl...

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>