iniscripts: s/tty0/tty in populate-volatiles.sh
/dev/tty0 is only present when a kernel is configured with
CONFIG_VT_CONSOLE, so the output redirection is incorrect when /dev/tty0
is not present.  Use /dev/tty since "it is a synonym for the controlling
terminal of a process, if any." (from tty(4) manual page).

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>
---
Another option would be /dev/console.  /dev/tty was chosen since it's
closer to the original code.