linux-openxt: blktap2: Avoid hang on close
A hang was observed in blktap_control_destroy_tap_ioctl() where
ring->n_pending was 1.  However, the tapdisk process had exited, so
ring->vma was NULL and the pending would never be serviced (and
n_pending never decremented).

Avoid the loop and wait_event when ring->vma is null.  Also break the
wait_event when ring->vma is null to avoid waiting for pending requests
that will never be serviced.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>