python-daemon: Convert to python3 and uprev
Version 1.5.5 doesn't support python3, so we have to uprev while
switching version.

Patch out the twine dependency.  It's not needed for using the library
(only uploading to pypi), and it pulls in too many dependencies.

For inherit pypi to generate correct SRC_URI & S, we must override
PYPI_PACKAGE to retain the "python-" prefix in the name.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>